anubis: install some config files
Anubis is configured through environment variables. Install a
default.env file and the default bot policies file to /etc/anubis.
AIUI, the same file is also compiled into the binary as a default,
but it is useful to have it if you want to make changes.

As for the env file, it would probably be more useful together with a
rc.d script running something like

set -a
source ${SYSCONFDIR}/anubis/default.env
set +a
anubis

or something.