py-netsnmp-cffi: fix definition of time types on NetBSD.
Apparently, we can't use preprocessor macros in the limited C
dialect python's ffi understands, so this is another way to work
around this problem: via a patch and substitution from the Makefile.
This fixes it at least for NetBSD, but non-Linux and non-NetBSD are
probably left out in the cold.  A better way to solve this would
be nice.

Bump PKGREVISION.