wip/input-leap: import input-leap-3.0.2 as net/input-leap
Input Leap is software that mimics the functionality of a KVM
switch, which historically would allow you to use a single keyboard
and mouse to control multiple computers by physically turning a
dial on the box to switch the machine you're controlling at any
given moment. Input Leap does this in software, allowing you to
tell it which machine to control by moving your mouse to the edge
of the screen, or by using a keypress to switch focus to a different
system.