rust187: Apply ugly workaround for Darwin/arm64.  Pull from main pkgsrc.
Ideally this should be resolved in the build process and only set the
library path at the specific points that it is required, but for now at
least it doesn't appear to have any harmful side-effects setting it for the
entire build, and makes the package build again.

Limit the blast radius of the other LD_LIBRARY_PATH settings to NetBSD, and
add a comment to explain why.

Original commit in main pkgsrc by jperkin@