flux2: Update to 2.6.4
Changes:
v2.6.4
Flux v2.6.4 is a patch release that comes with various fixes. Users are
encouraged to upgrade for the best experience.

v2.6.3
Flux v2.6.3 is a patch release that comes with various fixes. Users are
encouraged to upgrade for the best experience.

v2.6.2
Flux v2.6.2 is a patch release that comes with various fixes. Users are
encouraged to upgrade for the best experience.

v2.6.1
Flux v2.6.1 is a patch release that comes with various fixes. Users are
encouraged to upgrade for the best experience.

v2.6.0
Flux v2.6.0 is a feature release. Users are encouraged to upgrade for
the best experience.

For a compressive overview of new features and API changes included in
this release, please refer to the Announcing Flux 2.6 GA blog
post: <https://fluxcd.io/blog/2025/05/flux-v2.6.0/>.

Overview of the new features:

- General availability release for the Flux OCI Artifacts APIs and
  `flux artifact` commands
- Support for OCI digests pinning (ImagePolicy, ImageUpdateAutomation)
- Object-level workload identity authentication (OCIRepository,
  ImageRepository, Kustomization, Alert Provider)
- Cache registry credentials for cloud providers (OCIRepository,
  ImageRepository)
- Git HTTP/S Mutual TLS authentication (GitRepository,
  ImageUpdateAutomation)
- Support for sparse checkout (GitRepository)
- Support for GitHub App authentication (Alert Provider)
- Support for managed Identity authentication to Azure Event Hub (Alert
  Provider)
- Customize the ID of the Git commit status with CEL expressions (Alert
  Provider)
- `WaitForTermination` deletion policy (Kustomization)
- `DisableChartDigestTracking` feature gate (HelmRelease)

Upgrade procedure:
To upgrade the APIs, make sure the new CRDs and controllers are
deployed, and then change the manifests in Git:

1. Set  `apiVersion: source.toolkit.fluxcd.io/v1` in the YAML files
   that contain `OCIRepository` definitions.
2. Add an annotation `api.fluxcd.io/upgrade: "v2.6.0"` to the
   `OCIRepository` resources. (this is not required if Flux Operator is
   used for upgrade)
3. Commit, push, and reconcile the API version changes.

Bumping the APIs version in manifests can be done gradually.
It is advised to not delay this procedure as the deprecated versions
will be removed after 6 months.

v2.5.1
Flux v2.5.1 is a patch release which comes with various fixes. Users
are encouraged to upgrade for the best experience.

v2.5.0
Flux v2.5.0 is a feature release. Users are encouraged to upgrade for
the best experience.

For a compressive overview of new features and API changes included in
this release, please refer to the Announcing Flux 2.5 GA blog post:
<https://fluxcd.io/blog/2025/02/flux-v2.5.0/>.

Overview of the new features:

- Support for GitHub App authentication (`GitRepository` and
  `ImageUpdateAutomation` API)
- Custom Health Checks using CEL (`Kustomization` API)
- Fine-grained control of garbage collection (`Kustomization` API)
- Enable decryption of secrets generated by Kustomize components (`Kustomization` API)
- Support for custom event metadata from annotations (`Alert` API)
- Git commit status updates for Flux Kustomizations with OCIRepository
  sources (`Alert` API)
- Resource filtering using CEL for webhook receivers (`Receiver` API)
- Debug commands for Flux Kustomizations and HelmReleases (Flux CLI)

Upgrade procedure:
There are no new API versions in this release, so no changes are
required in the YAML manifests containing Flux resources.