Forgejo: update to 13.0.1
Content moderation is one of the main roadblocks for federation. A
first step is now available to users and admins. It was completed
while ActivityPub related pull requests were worked on in parallel and
merged.

Security was improved under the hood for Forgejo Actions secrets now
using a more secure module introduced in 2024 and already in use for
TOTP secrets since Forgejo v10. In addition it is now possible to
require 2FA instance wide.

Forgejo Actions usability was improved with access to all run attempts
of each workflow: they were stored but not available via the web
UI. The workflow files are now statically checked, for early detection
of errors such as the usage of a wrong context in an expression like
${{ forgejo.server_url }}.