flux2: Update to 2.7.5
Changes:
v2.7.5
Flux v2.7.5 is a patch release that comes with fixes to
helm-controller. Users are encouraged to upgrade for the best
experience.

v2.7.4
Flux v2.7.4 is a patch release that comes with various fixes. Users are
encouraged to upgrade for the best experience.

v2.7.3
Flux v2.7.3 is a patch release that comes with various fixes. Users are
encouraged to upgrade for the best experience.

v2.7.2
Flux v2.7.2 is a patch release that comes with security fixes. Users
are encouraged to upgrade for the best experience.

v2.7.1
Flux v2.7.1 is a patch release that comes with various improvements and
fixes. Users are encouraged to upgrade for the best experience.

v2.7.0
Flux v2.7.0 is a feature release. Users are encouraged to upgrade for
the best experience.

For a compressive overview of new features and API changes included in
this release, please refer to the [Announcing Flux 2.7 GA blog
post at <https://fluxcd.io/blog/2025/09/flux-v2.7.0/>.

Upgrade procedure
The Flux APIs `v1beta1` and `v2beta1` (deprecated in 2023) have reached
end-of-life and have been removed from the CRDs.

Unless you are using Flux Operator to deploy the Flux controllers, you
must run the `flux migrate` command on clusters before upgrading.

For more details, please refer to the Flux v2.7 upgrade guide at
<https://fluxcd.io/blog/2025/09/flux-v2.7.0/#upgrade-procedure>.